pyproject.toml
[project]
name = "python-gui-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.0",
    "numpy>=2.2.1",
    "pyedifice>=2.9.1",
    "pyside6-essentials>=6.8.1,!=6.8.1.1",
    "watchdog>=6.0.0",
]

[tool.uv.pip]
generate-hashes = true